// Project/InternalAuditSystem.Web/ClientApp/src/app/swaggerapi/AngularFiles/model/strategicAnalysisTeamAC.ts
/**
 * My Title
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { BaseModelAC } from './baseModelAC';
import { StrategicAnalysisAC } from './strategicAnalysisAC';
import { UserAC } from './userAC';

export interface StrategicAnalysisTeamAC extends BaseModelAC { 
    strategicAnalysisId?: string;
    strategicAnalysis?: StrategicAnalysisAC;
    userId?: string;
    user?: UserAC;
    name?: string;
    designation?: string;
    emailId?: string;
    auditableEntityName?: string;
}