// Project/InternalAuditSystem.Web/ClientApp/src/app/swaggerapi/AngularFiles/model/userResponseAC.ts
/**
 * My Title
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { BaseModelAC } from './baseModelAC';
import { OptionAC } from './optionAC';
import { QuestionAC } from './questionAC';
import { SamplingResponseStatus } from './samplingResponseStatus';
import { StrategicAnalysisStatus } from './strategicAnalysisStatus';
import { UserAC } from './userAC';

export interface UserResponseAC extends BaseModelAC { 
    questionId?: string;
    question?: QuestionAC;
    userId?: string;
    user?: UserAC;
    parentFileUploadUserResponseId?: string;
    optionIds?: Array<string>;
    options?: Array<OptionAC>;
    other?: string;
    representationNumber: number;
    answerText?: string;
    strategicAnalysisId?: string;
    multipleChoiceQuestionId?: string;
    checkboxQuestionId?: string;
    dropdownQuestionId?: string;
    files?: Array<Blob>;
    checkboxOptions?: Array<string>;
    mcqOption?: string;
    selectedDropdownOption?: string;
    userResponseStatus: StrategicAnalysisStatus;
    samplingResponseStatus?: SamplingResponseStatus;
    userResponseStatusString?: string;
    isEmailAttachements: boolean;
    riskControlMatrixId?: string;
    auditableEntityId?: string;
    auditableEntityName?: string;
}